refactor(expand): separate string buffer from field construction
