+++ test/unit_test/expand/Makefile
@@ -21,8 +21,9 @@ ROOT_SRCS:=parser/parser.c \
 	lexer/lexer_internal/lex_word_util.c \
 	lexer/lexer_internal/lexer_utils.c \
 	expand/expand.c \
-	expand/expand_internal/expand_buf.c \
+	expand/expand_internal/expand_buffer.c \
 	expand/expand_internal/expand_cmd.c \
+	expand/expand_internal/expand_fields.c \
 	expand/expand_internal/expand_glob.c \
 	expand/expand_internal/expand_heredoc.c \
 	expand/expand_internal/expand_param.c \
